refactor(bridge): fold the topic-split files back into their four real layers
Every one of internal/bridge's 22 source files was created by one commit
(00bae323, "extract bridge to internal/bridge") — so the interact_-style
filename prefixes are that commit's naming taste, not co-change evidence.
The reference graph disagrees with them in four places, and this follows the
graph:

- bridge_forward.go -> bridge.go. The read loop's only job is to hand each
  line to bridgeForwardRequest, and the forwarder's respawn+retry is what
  makes the loop survive a daemon death. One unit, two files.
- bridge_transport_{helpers,errors,restart}.go -> bridge_transport.go. All
  three exist solely to serve StdioToHTTPFast; nothing else calls into them.
- bridge_fastpath_telemetry.go -> bridge_fastpath.go. The counters have
  exactly one producer (handleFastPath); the split implied a reusable
  telemetry subsystem that was never reusable.
- bridge_startup{,_lock,_status}.go -> bridge_startup.go. Every timing knob
  in the old bridge_startup.go is read by at least two of the other three, so
  the startup policy could not be read in one place.

bridge_startup_state.go is deliberately NOT folded in: the daemon state
machine (respawn throttling, ready/failed signalling) is the one piece here
that is independently testable, and the recent crash-loop/backoff fixes live
there.

Two source-contract tests were pinned to filenames (parser.ParseFile of
"bridge_forward.go"/"bridge.go") and so were one rename away from being
silently unenforceable. They now locate their subject by symbol via a shared
findFuncDecl helper.

cmd/browser-agent/internal/bridge: 17 -> 10 source files (baseline 22 -> out).
Package coverage 71.5% -> 74.6%.
